Fixed issue #3 with minor refactor.
The sys_profile method was refactored due to the attribute having the
same name. This also adds a conditional at install time to respect the
value.

Currently the tests for this are written as a duplicate serverspec
suite with a negative test for the sys_profile link. This is not ideal
as it is time consuming to test this. This will need to be refactored
into rspec at some point.