Fixed catkin workspace resource.
The catkin resource now fully creates and initializes the catkin
workspace. There is one known issue which seems to be inconsequential
in the converge where The syntax error message is displayed.

 "sh: 1: Syntax error: ";" unexpected"

I am not entirely sure why the error occurs, as it is proper shell
syntax for executing two commands. Thusfar it seems to have no negative
impact.