Increase DigitalOcean VM size to 1GB.
The ros_comm package would exhaust all available memory on a 512MB
instance while compiling the C++ components, resulting in travis
failures. This increases the RAM to 1GB which resolves the build
failure.

As a side note, ros_comm may not be the best package for this testing
afterall given the time it takes to compile. I will likely revisit this
again once I find a more ideal canidate.